Regrouped outline providers and their respective node class.
