--- src/leoGotoNode.ts
@@ -1,101 +1,0 @@
-import * as vscode from "vscode";
-import { Constants } from "./constants";
-import { LeoIntegration } from "./leoIntegration";
-import { LeoGoto, TGotoTypes } from "./types";
-import * as utils from "./utils";
-
-/**
- * * Opened Leo documents tree view node item implementation for usage in a TreeDataProvider
- */
-export class LeoGotoNode extends vscode.TreeItem {
-
-    // Context string is checked in package.json with 'when' clauses
-    public entryType: TGotoTypes;
-    private _id: string;
-    private _description: string | boolean;
-    private _headline: string;
-    private _iconIndex: number; // default to tag
-    private _leoIntegration: LeoIntegration;
-    public key: number; // id from python
-
-    constructor(
-        p_leoIntegration: LeoIntegration,
-        p_gotoEntry: LeoGoto,
-        p_navOptions: { isTag: boolean, showParents: boolean },
-
-    ) {
-        let w_spacing = "";
-        if (p_navOptions.showParents && !p_navOptions.isTag) {
-            w_spacing = "    ";
-        }
-        let w_label = "";
-        if (["tag", "headline"].includes(p_gotoEntry.t)) {
-            w_label = w_spacing + p_gotoEntry.h;
-        }
-        super(w_label);
-
-        // Setup this instance
-        this._leoIntegration = p_leoIntegration;
-        this._id = utils.getUniqueId();
-        this.entryType = p_gotoEntry.t;
-        this.key = p_gotoEntry.key;
-        this._headline = p_gotoEntry.h.trim();
-
-        this._description = false;
-        if (this.entryType === 'body') {
-            this._iconIndex = 2;
-            if (p_navOptions.showParents) {
-                this._description = "    " + this._headline;
-            } else {
-                this._description = "  " + this._headline;
-            }
-        } else if (this.entryType === 'parent') {
-            this._iconIndex = 0;
-            this._description = this._headline.trim();
-        } else if (this.entryType === 'generic') {
-            this._iconIndex = 4;
-            this._description = this._headline;
-        } else if (this.entryType === 'headline') {
-            this._iconIndex = 1;
-        } else {
-            this._iconIndex = 3; // tag
-        }
-
-        this.command = {
-            command: Constants.COMMANDS.GOTO_NAV_ENTRY,
-            title: '',
-            arguments: [this]
-        };
-
-    }
-
-    // @ts-ignore
-    public get tooltip(): string {
-        if (this.entryType !== "generic") {
-            return this.entryType.charAt(0).toUpperCase() + this.entryType.slice(1);
-        }
-        return this._headline;
-    }
-
-    // @ts-ignore
-    public get description(): string | boolean {
-        return this._description;
-    }
-
-    // @ts-ignore
-    public get iconPath(): Icon | vscode.ThemeIcon | string {
-        if (this._iconIndex < 4) {
-            return this._leoIntegration.gotoIcons[this._iconIndex];
-        }
-        // else return undefined for generic text without icon
-        return undefined;
-    }
-
-    // @ts-ignore
-    public get id(): string {
-        // Add prefix and suffix salt to numeric index to prevent accidental duplicates
-        // Should be unique when refreshed
-        return "g" + this._id + "o";
-    }
-
-}
