+++ src/leoUndos.ts
@@ -1,9 +1,6 @@
 import * as vscode from "vscode";
 import { LeoIntegration } from "./leoIntegration";
-import { LeoUndoNode } from "./leoUndoNode";
-import { ProviderResult } from "vscode";
 import { Constants } from "./constants";
-import { LeoUndo } from "./types";
 
 // leo undo provider class
 export class LeoUndosProvider implements vscode.TreeDataProvider<LeoUndoNode> {
@@ -143,4 +140,24 @@ export class LeoUndosProvider implements vscode.TreeDataProvider<LeoUndoNode> {
         // item.tooltip = "TODO Undo Tooltip";
         return item;
     }
+}
+
+/**
+ * * Opened Leo documents tree view node item implementation for usage in a TreeDataProvider
+ */
+export class LeoUndoNode extends vscode.TreeItem {
+
+    // Context string is checked in package.json with 'when' clauses
+    public contextValue: string = "leoUndoNode";
+
+    constructor(
+        public label: string,
+        public description: string,
+        public id: string,
+        // public contextValue: string,
+        public beadIndex: number
+    ) {
+        super(label);
+    }
+
 }
